+++ cmd_version.go
@@ -5,8 +5,8 @@ import (
 )
 
 var CmdVersion = &Cmd{
-	Name:    "version",
-	Private: true,
+	Name: "version",
+	Desc: "prints the version",
 	Fn: func(env Env, args []string) error {
 		fmt.Println(VERSION)
 		return nil