+++ config.go
@@ -6,15 +6,32 @@ import (
 	"os/exec"
 	"path/filepath"
 	"strings"
+
+	toml "github.com/BurntSushi/toml"
 )
 
 type Config struct {
-	Env      Env
-	WorkDir  string // Current directory
-	ConfDir  string
-	SelfPath string
-	BashPath string
-	RCDir    string
+	Env             Env
+	WorkDir         string // Current directory
+	ConfDir         string
+	SelfPath        string
+	BashPath        string
+	RCDir           string
+	TomlPath        string
+	DisableStdin    bool
+	WhitelistPrefix []string
+	WhitelistExact  map[string]bool
+}
+
+type tomlConfig struct {
+	Whitelist    whitelist `toml:"whitelist"`
+	BashPath     string    `toml:"bash_path"`
+	DisableStdin bool      `toml:"disable_stdin"`
+}
+
+type whitelist struct {
+	Prefix []string
+	Exact  []string
 }
 
 func LoadConfig(env Env) (config *Config, err error) {
@@ -43,16 +60,6 @@ func LoadConfig(env Env) (config *Config, err error) {
 	exePath = strings.Replace(exePath, "\\", "/", -1)
 	config.SelfPath = exePath
 
-	config.BashPath = env[DIRENV_BASH]
-	if config.BashPath == "" {
-		if bashPath != "" {
-			config.BashPath = bashPath
-		} else if config.BashPath, err = exec.LookPath("bash"); err != nil {
-			err = fmt.Errorf("Can't find bash: %q", err)
-			return
-		}
-	}
-
 	if config.WorkDir, err = os.Getwd(); err != nil {
 		err = fmt.Errorf("LoadConfig() Getwd failed: %q", err)
 		return
@@ -61,6 +68,45 @@ func LoadConfig(env Env) (config *Config, err error) {
 	config.RCDir = env[DIRENV_DIR]
 	if len(config.RCDir) > 0 && config.RCDir[0:1] == "-" {
 		config.RCDir = config.RCDir[1:]
+	}
+
+	config.WhitelistPrefix = make([]string, 0)
+	config.WhitelistExact = make(map[string]bool)
+
+	// Load the config.toml
+	config.TomlPath = filepath.Join(config.ConfDir, "config.toml")
+	if _, statErr := os.Stat(config.TomlPath); statErr == nil {
+		var tomlConf tomlConfig
+		if _, err = toml.DecodeFile(config.TomlPath, &tomlConf); err != nil {
+			err = fmt.Errorf("LoadConfig() failed to parse config.toml: %q", err)
+			return
+		}
+
+		for _, prefix := range tomlConf.Whitelist.Prefix {
+			config.WhitelistPrefix = append(config.WhitelistPrefix, prefix)
+		}
+
+		for _, path := range tomlConf.Whitelist.Exact {
+			if !strings.HasSuffix(path, "/.envrc") {
+				path = filepath.Join(path, ".envrc")
+			}
+
+			config.WhitelistExact[path] = true
+		}
+
+		config.DisableStdin = tomlConf.DisableStdin
+		config.BashPath = tomlConf.BashPath
+	}
+
+	if config.BashPath == "" {
+		if env[DIRENV_BASH] != "" {
+			config.BashPath = env[DIRENV_BASH]
+		} else if bashPath != "" {
+			config.BashPath = bashPath
+		} else if config.BashPath, err = exec.LookPath("bash"); err != nil {
+			err = fmt.Errorf("Can't find bash: %q", err)
+			return
+		}
 	}
 
 	return
@@ -79,11 +125,11 @@ func (self *Config) LoadedRC() *RC {
 
 	times_string := self.Env[DIRENV_WATCHES]
 
-	return RCFromEnv(rcPath, times_string)
+	return RCFromEnv(rcPath, times_string, self)
 }
 
 func (self *Config) FindRC() *RC {
-	return FindRC(self.WorkDir, self.AllowDir())
+	return FindRC(self.WorkDir, self)
 }
 
 func (self *Config) EnvDiff() (*EnvDiff, error) {