+++ stdlib.go
@@ -5,10 +5,11 @@ const STDLIB = "#!bash\n" +
 	"# These are the commands available in an .envrc context\n" +
 	"#\n" +
 	"set -e\n" +
-	"direnv=\"%s\"\n" +
+	"# NOTE: don't touch the RHS, it gets replaced at runtime\n" +
+	"direnv=\"$(which direnv)\"\n" +
 	"\n" +
 	"# Config, change in the direnvrc\n" +
-	"DIRENV_LOG_FORMAT=\"${DIRENV_LOG_FORMAT-direnv: %%s}\"\n" +
+	"DIRENV_LOG_FORMAT=\"${DIRENV_LOG_FORMAT-direnv: %s}\"\n" +
 	"\n" +
 	"# Usage: direnv_layout_dir\n" +
 	"#\n" +
@@ -429,6 +430,57 @@ const STDLIB = "#!bash\n" +
 	"  layout_python python3 \"$@\"\n" +
 	"}\n" +
 	"\n" +
+	"# Usage: layout anaconda <enviroment_name> [<conda_exe>]\n" +
+	"#\n" +
+	"# Activates anaconda for the named environment. If the environment\n" +
+	"# hasn't been created, it will be using the environment.yml file in\n" +
+	"# the current directory. <conda_exe> is optional and will default to\n" +
+	"# the one found in the system environment.\n" +
+	"#\n" +
+	"layout_anaconda() {\n" +
+	"  local env_name=$1\n" +
+	"  local conda\n" +
+	"  if [[ $# -gt 1 ]]; then\n" +
+	"    conda=${2}\n" +
+	"  else\n" +
+	"    conda=$(command -v conda)\n" +
+	"  fi\n" +
+	"  PATH_add $(dirname \"$conda\")\n" +
+	"  local env_loc=$(\"$conda\" env list | grep -- \"$env_name\")\n" +
+	"  if [[ ! \"$env_loc\" == $env_name*$env_name ]]; then\n" +
+	"    if [[ -e environment.yml ]]; then\n" +
+	"      log_status \"creating conda enviroment\"\n" +
+	"      \"$conda\" env create\n" +
+	"    else\n" +
+	"      log_error \"Could not find environment.yml\"\n" +
+	"      return 1\n" +
+	"    fi\n" +
+	"  fi\n" +
+	"\n" +
+	"  source activate \"$env_name\"\n" +
+	"}\n" +
+	"\n" +
+	"# Usage: layout pipenv\n" +
+	"#\n" +
+	"# Similar to layout_python, but uses Pipenv to build a\n" +
+	"# virtualenv from the Pipfile located in the same directory.\n" +
+	"#\n" +
+	"layout_pipenv() {\n" +
+	"  if [[ ! -f Pipfile ]]; then\n" +
+	"    log_error 'No Pipfile found.  Use `pipenv` to create a Pipfile first.'\n" +
+	"    exit 2\n" +
+	"  fi\n" +
+	"\n" +
+	"  local VENV=$(pipenv --bare --venv 2>/dev/null)\n" +
+	"  if [[ -z $VENV || ! -d $VENV ]]; then\n" +
+	"    pipenv install --dev\n" +
+	"  fi\n" +
+	"\n" +
+	"  export VIRTUAL_ENV=$(pipenv --venv)\n" +
+	"  PATH_add \"$VIRTUAL_ENV/bin\"\n" +
+	"  export PIPENV_ACTIVE=1\n" +
+	"}\n" +
+	"\n" +
 	"# Usage: layout ruby\n" +
 	"#\n" +
 	"# Sets the GEM_HOME environment variable to \"$(direnv_layout_dir)/ruby/RUBY_VERSION\".\n" +
@@ -586,10 +638,19 @@ const STDLIB = "#!bash\n" +
 	"# (e.g `use nix -p ocaml`).\n" +
 	"#\n" +
 	"use_nix() {\n" +
+	"  local orig_IN_NIX_SHELL=\"$IN_NIX_SHELL\"\n" +
+	"\n" +
 	"  direnv_load nix-shell --show-trace \"$@\" --run 'direnv dump'\n" +
 	"  if [[ $# = 0 ]]; then\n" +
 	"    watch_file default.nix\n" +
 	"    watch_file shell.nix\n" +
+	"  fi\n" +
+	"\n" +
+	"  # Don't change the IN_NIX_SHELL env var\n" +
+	"  if [[ -z $orig_IN_NIX_SHELL ]]; then\n" +
+	"    unset IN_NIX_SHELL\n" +
+	"  else\n" +
+	"    export IN_NIX_SHELL=\"$orig_IN_NIX_SHELL\"\n" +
 	"  fi\n" +
 	"}\n" +
 	"\n" +